Deuxieme Correction

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: fleriche <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 14:17:09 by fleriche          #+#    #+#             */
-/*   Updated: 2022/11/14 11:51:35 by fleriche         ###   ########.fr       */
+/*   Updated: 2022/12/12 09:33:48 by fleriche         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	a;
