+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: fleriche <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 14:30:13 by fleriche          #+#    #+#             */
-/*   Updated: 2022/11/22 10:40:44 by fleriche         ###   ########.fr       */
+/*   Updated: 2022/12/12 09:38:45 by fleriche         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int		ft_isascii(int c);
 int		ft_isdigit(int c);
 int		ft_isprint(int c);
 void	*ft_memset(void *b, int c, size_t len);
-int		ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 void	ft_bzero(void *s, size_t n);
 void	*ft_memmove(void *dst, const void *src, size_t len);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
